[COMMS-889] Refactor to match current file structure
https://community.openproject.org/wp/COMMS-889
