+++ frontend/src/turbo/turbo-global-listeners.spec.ts
@@ -27,10 +27,9 @@
 //++
 
 import {
-  afterEach, beforeEach, describe, expect, it, vi,
+  afterEach, beforeEach, describe, expect, it,
 } from 'vitest';
-import * as Turbo from '@hotwired/turbo';
-import { addTurboGlobalListeners, canonicalizeWorkPackageIdInUrl } from './turbo-global-listeners';
+import { addTurboGlobalListeners } from './turbo-global-listeners';
 
 describe('addTurboGlobalListeners — OPCE custom element morph guard', () => {
   let controller:AbortController;
@@ -68,81 +67,5 @@ describe('addTurboGlobalListeners — OPCE custom element morph guard', () => {
     const notCancelled = beforeMorph(element);
 
     expect(notCancelled).toBe(true);
-  });
-});
-
-describe('canonicalizeWorkPackageIdInUrl', () => {
-  let historyReplaceSpy:ReturnType<typeof vi.spyOn>;
-  let canonicalLink:HTMLLinkElement | null = null;
-
-  function setCanonical(href:string) {
-    const link = document.createElement('link');
-    link.rel = 'canonical';
-    link.href = href;
-    document.head.appendChild(link);
-    canonicalLink = link;
-  }
-
-  beforeEach(() => {
-    /* eslint-disable @typescript-eslint/no-empty-function */
-    historyReplaceSpy = vi.spyOn(Turbo.session.history, 'replace').mockImplementation(() => {});
-    /* eslint-enable @typescript-eslint/no-empty-function */
-  });
-
-  afterEach(() => {
-    // Deterministic per-test cleanup: restore the spy, drop the injected canonical
-    // link and reset the URL rather than relying on the next test's setup.
-    vi.restoreAllMocks();
-    canonicalLink?.remove();
-    canonicalLink = null;
-    window.history.replaceState({}, '', '/');
-  });
-
-  it('rewrites old id to canonical id', () => {
-    window.history.pushState({}, '', '/projects/old-proj/work_packages/OLD-42/activity');
-    setCanonical('http://localhost/projects/new-proj/work_packages/NEW-42/activity');
-
-    canonicalizeWorkPackageIdInUrl();
-
-    const expectedUrl = new URL('/projects/old-proj/work_packages/NEW-42/activity', window.location.origin);
-    expect(historyReplaceSpy).toHaveBeenCalledWith(expectedUrl, Turbo.session.history.restorationIdentifier);
-    expect((Turbo.session as unknown as { view:{ lastRenderedLocation:URL } }).view.lastRenderedLocation.href)
-      .toBe(expectedUrl.href);
-  });
-
-  it('does not rewrite when url already matches canonical', () => {
-    window.history.pushState({}, '', '/projects/demo/work_packages/DEMO-42');
-    setCanonical('http://localhost/projects/demo/work_packages/DEMO-42');
-
-    canonicalizeWorkPackageIdInUrl();
-
-    expect(historyReplaceSpy).not.toHaveBeenCalled();
-  });
-
-  it('does not rewrite when no canonical link is present', () => {
-    window.history.pushState({}, '', '/projects/demo/work_packages/42');
-
-    canonicalizeWorkPackageIdInUrl();
-
-    expect(historyReplaceSpy).not.toHaveBeenCalled();
-  });
-
-  it('does not rewrite when url does not contain /work_packages/', () => {
-    window.history.pushState({}, '', '/projects/demo/boards');
-    setCanonical('http://localhost/projects/demo/boards');
-
-    canonicalizeWorkPackageIdInUrl();
-
-    expect(historyReplaceSpy).not.toHaveBeenCalled();
-  });
-
-  it('preserves query string and hash when rewriting the id', () => {
-    window.history.pushState({}, '', '/projects/demo/work_packages/13?focus=description#comment-5');
-    setCanonical('http://localhost/projects/demo/work_packages/DEMO-42');
-
-    canonicalizeWorkPackageIdInUrl();
-
-    const expectedUrl = new URL('/projects/demo/work_packages/DEMO-42?focus=description#comment-5', window.location.origin);
-    expect(historyReplaceSpy).toHaveBeenCalledWith(expectedUrl, Turbo.session.history.restorationIdentifier);
   });
 });
