+++ spec/features/work_packages/canonical_url_rewrite_spec.rb
@@ -1,5 +1,33 @@
 # frozen_string_literal: true
 
+# -- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+# ++
+
 require "spec_helper"
 
 RSpec.describe "Work package canonical URL rewrite",
@@ -34,6 +62,21 @@ RSpec.describe "Work package canonical URL rewrite",
   end
 
   describe "after Turbo Drive navigation (turbo:render)" do
+    it "rewrites a numeric work package ID to the canonical semantic ID" do
+      visit project_path(project)
+      expect(page).to have_current_path(project_path(project))
+
+      # No real UI link can produce a Turbo Drive visit to a work package URL:
+      # user-content links are rendered with target="_top", which Turbo 8 never
+      # intercepts (findLinkFromClickTarget ignores links with target != _self).
+      # Trigger the visit directly to exercise the turbo:render code path.
+      wait_for_turbo { page.execute_script("Turbo.visit(#{numeric_path.to_json})") }
+
+      expect(page).to have_current_path(semantic_path)
+    end
+  end
+
+  describe "following a user-content link (full page load)" do
     let!(:wiki_page) do
       create(:wiki_page,
              wiki: project.wiki,
@@ -45,9 +88,11 @@ RSpec.describe "Work package canonical URL rewrite",
       visit project_wiki_path(project, wiki_page)
       expect(page).to have_css(".wiki-content")
 
-      wait_for_turbo do
-        within(".wiki-content") { click_link "the work package" }
-      end
+      # User-content links carry target="_top" (format_text), which Turbo does not
+      # intercept — the click is a full page load, so the rewrite runs on
+      # DOMContentLoaded and the self-waiting matcher below covers it. wait_for_turbo
+      # cannot be used across a full load (its listener registry does not survive).
+      within(".wiki-content") { click_link "the work package" }
 
       expect(page).to have_current_path(semantic_path)
       full_screen.ensure_page_loaded
